server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.23
psycopg[binary]>=3.2.11
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.0.0
PyMuPDF>=1.23.0
sse-starlette>=1.8.0
httpx>=0.25.0
